google_gemini_genai/simple.py
from google import genai
from google.genai import types

# The client gets the API key from the environment variable `GEMINI_API_KEY`.
client = genai.Client()

response = client.models.generate_content(
    model="gemini-3-flash-preview",
    config=types.GenerateContentConfig(
        system_instruction="You are a marketing expert."
    ),
    contents="What would be a good company name for a company that makes colorful socks?",
)
print(response.text)
