google_gemini_genai/pyproject.toml
[project]
name = "google-gemini-genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.66.0",
]
